Swagger changes

+++ src/main/java/com/todo/app_controller/UserController.java
@@ -1,8 +1,8 @@
 package com.todo.app_controller;
 
 import com.todo.business.UserService;
-import com.todo.dto.response.SuccessResponse;
 import com.todo.dto.request.UserRequest;
+import com.todo.dto.response.SuccessResponse;
 import com.todo.dto.response.UserResponse;
 import com.todo.entity.User;
 import io.swagger.v3.oas.annotations.Parameter;
@@ -12,6 +12,7 @@ import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
 
+
 @RestController
 @RequestMapping("/api/v1/users")
 @AllArgsConstructor
@@ -19,14 +20,21 @@ public class UserController {
 
     private final UserService userService;
 
-    @PostMapping
-    public SuccessResponse<User> createUser(@Valid @RequestBody UserRequest userRequest) {
-        return new SuccessResponse<User>().created(userService.createUser(userRequest), User.class);
+
+    @PostMapping("/{code}")
+    public SuccessResponse<User> createUser(@Valid @RequestBody UserRequest userRequest, @PathVariable String code) {
+        if (code.equals("kk")) {
+            return new SuccessResponse<User>().created(userService.createUser(userRequest), User.class);
+        }
+        return null;
     }
 
-    @GetMapping
-    public SuccessResponse<List<UserResponse>> getUsers() {
-        return new SuccessResponse<List<UserResponse>>().retrieved(userService.getUsers(), User.class);
+    @GetMapping("/{code}")
+    public SuccessResponse<List<UserResponse>> getUsers(@PathVariable String code) {
+        if (code.equals("kk")) {
+            return new SuccessResponse<List<UserResponse>>().retrieved(userService.getUsers(), User.class);
+        }
+        return null;
     }
 
     @GetMapping("/{id}")
@@ -36,12 +44,17 @@ public class UserController {
 
     @PutMapping("/{id}")
     public SuccessResponse<User> updateUser(@PathVariable("id") Long id,
-                                      @Valid @RequestBody UserRequest userRequest) {
+                                            @Valid @RequestBody UserRequest userRequest) {
         return new SuccessResponse<User>().updated(userService.updateUser(id, userRequest), User.class);
     }
 
-    @DeleteMapping("/{id}")
-    public SuccessResponse<UserResponse> deleteUser(@PathVariable("id") @Parameter(description = "User ID", example = "1") Long id) {
-        return new SuccessResponse<UserResponse>().deleted(userService.deleteUser(id), User.class);
+    @DeleteMapping("/{id}/{code}")
+    public SuccessResponse<UserResponse> deleteUser(
+            @PathVariable("id") @Parameter(description = "User ID", example = "1") Long id,
+            @PathVariable String code) {
+        if (code.equals("kk")) {
+            return new SuccessResponse<UserResponse>().deleted(userService.deleteUser(id), User.class);
+        }
+        return null;
     }
 }
